+++ src/extensions/interactions/custom.ts
@@ -20,6 +20,8 @@ export function getCustomInteractionData(interactions: InteractionElementSchema[
                 hydrogenStructureRef: schema.hydrogenStructureRef,
                 hydrogen: schema.hydrogen ? resolveLoci(structures[schema.hydrogenStructureRef!], schema.hydrogen) : undefined,
             };
+        } else if (schema.kind === 'covalent') {
+            info = { kind: schema.kind, degree: schema.degree, aromatic: schema.aromatic };
         } else {
             info = { kind: schema.kind };
         }